+++ extensions/omni_httpc/migrate/3_http_execute.sql
@@ -14,6 +14,6 @@ create domain valid_http_execute_options as http_execute_options
             ((value).http2_ratio + (value).http3_ratio) <= 100
         );
 
-/*{% include "../src/http_execute_options.sql" %}*/
+/*{% include "../src/http_execute_options_prev.sql" %}*/
 /*{% include "../src/http_execute.sql" %}*/
 /*{% include "../src/http_execute_with_options.sql" %}*/