+++ extensions/omni_httpc/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - TBD
+
+### Addded
+
+* Support for custom CA certificates [#676](https://github.com/omnigres/omnigres/pull/676)
+* Support for custom client certificates [#676](https://github.com/omnigres/omnigres/pull/676)
+* An option to ignore self-signed certificates [#676](https://github.com/omnigres/omnigres/pull/676)
+
 ## [0.1.2] - 2024-09-27
 
 ### Added
@@ -21,10 +29,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Initial release following a few months of iterative development.
 
-[Unreleased]: https://github.com/omnigres/omnigres/commits/next/omni_httpd
+[Unreleased]: https://github.com/omnigres/omnigres/commits/next/omni_httpc
 
 [0.1.0]: [https://github.com/omnigres/omnigres/pull/511]
 
 [0.1.1]: [https://github.com/omnigres/omnigres/pull/578]
 
 [0.1.2]: [https://github.com/omnigres/omnigres/pull/650]
+
+[0.1.3]: [https://github.com/omnigres/omnigres/pull/676]